# Optimization: Use either monotonic or wall-clock to measure command execution time, to regain up to 4% execution time