In #7491 (part of redis 6.2), we started using the monotonic timer instead of mstime to measure command execution time for stats, apparently this meant sampling the clock 3 times per command rather than two (wince we also need the wall-clock time).
In some cases this causes a significant overhead.

This PR fixes that by avoiding the use of monotonic timer, except for the cases were we know it should be extremely fast.
This PR also adds a new INFO field called `monotonic_clock` that shows which clock redis is using.

background
====

We saw in #10460 that updateCachedTime took ~1.5% of CPU cycles to get the time once again, when we already keep track of time within call() for the command latency tracking. 
Given we need the UNIX timestamp for updateCachedTime() call we might as well measure time once before the command logic by using ustime() instead of having to use ustime() + elapsedStart(). 

to reproduce:
```
# populate data
memtier_benchmark --ratio 1:0  -d 1000 --key-pattern P:P --key-maximum 1000000  --key-minimum 1 --pipeline 15 --hide-histogram

# benchmark
memtier_benchmark --ratio 0:1 --test-time 60  -d 1000  --key-maximum 1000000  --key-minimum 1 --pipeline 15 --hide-histogram
```

## before profile and results:
```
ALL STATS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets            0.00          ---          ---             ---             ---             ---             ---         0.00 
Gets       676114.65    676114.65         0.00         4.43638         3.74300         6.94300         7.48700    688585.47 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     676114.65    676114.65         0.00         4.43638         3.74300         6.94300         7.48700    688585.47 

```
![image](https://user-images.githubusercontent.com/5832149/161114458-86f07acb-666c-4e64-97c1-9d15cf6c13a5.png)

## after results:

ustime:
```
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets            0.00          ---          ---             ---             ---             ---             ---         0.00 
Gets       689547.70    689547.70         0.00         4.34994         3.74300         7.29500         7.67900    702266.28 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     689547.70    689547.70         0.00         4.34994         3.74300         7.29500         7.67900    702266.28 
```

hw clock (monotonic clock: X86 TSC @ 2100 ticks/us) :
```
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets            0.00          ---          ---             ---             ---             ---             ---         0.00 
Gets       691869.31    691869.31         0.00         4.33539         3.71100         7.23100         7.58300    704630.70 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     691869.31    691869.31         0.00         4.33539         3.71100         7.23100         7.58300    704630.70
```

this means we get ~2% improvement by default @oranagra. the HW clock reduces the overhead of time tracking by 0.3%.