@filipecosta90 2022-04-05
> so you're saying this regression is from #7491 (part of redis 6.3), which started using the monotonic timer instead of mstime to measure command execution time stats?

yep. given we really need to the `ustime()` call, using the monotonic time tracking was a double effort.