@JimB123 2022-04-07
Let me see if I understand the conversation so far:
- We need to call `ustime()` because there is a cached timestamp which matches the wall-clock.  This is set at the beginning of every command and is used to support various functions which need this recent wall-clock time.
- We also need to time the execution of the current command by measuring time before/after invoking the command.  Measuring duration is something that we'd normally perform using the monotonic clock (not wall-clock).

The logic is something like this:
```
set cached_ustime = ustime()   // wall clock
...
set command_start = getMonotonic()
execute command
set command_duration = getMonotonic() - command_start
```
This logic is correct.  When the HW clock is used for monotonic, `getMonotonic()` is essentially free and the total cost is basically 1 call to `ustime()`.

However, if the HW clock is disabled, `getMonotonic()` is implemented in a similar way to calling `ustime()`.  This results in essentially 3 calls to `ustime()`.  We recognize that we could reduce this to 2 calls to `ustime()` by combining the caching of `cached_ustime` with the gathering of `command_start` from monotonic.

This would result in:
- If the HW clock is enabled, a performance reduction as 1 call to `ustime()` would become 2 calls.
- If the HW clock is disabled, a performance improvement as 3 calls to `ustime()` would become 2 calls.

If I understand correctly, the proposed change would choose to improve performance for some users at the expense of other users.

I think that the only reasonable way to make this optimization would be to actually check the HW flag with code something like this:

```
set cached_ustime = ustime()   // wall clock
...
#ifdef USE_PROCESSOR_CLOCK
    set command_start = getMonotonic()
#endif
execute command
#ifdef USE_PROCESSOR_CLOCK
    set command_duration = getMonotonic() - command_start
#else
    set command_duration = ustime() - cached_ustime
#endif
```

This is not ideal.  It's a bit ugly, and `USE_PROCESSOR_CLOCK` define would have to be relocated (into `config.h`?).

More ideally, we would convert most of Redis to use the monotonic clock rather than needing a cached wall-clock time.  If all of the code using the cached time was converted to use the monotonic clock, we wouldn't need any special logic here.  (But this is a large task and unrealistic for a single update.)