@JimB123 2022-04-07
> i don't think we can / want to convert all of redis to use monotonic clock. there are things that specifically require wall clock. all the places the measure time-delta and latencies should, use monotonic clock, but things like expiry (specifically EXPIREAT), and SLOWLOG timestamps must use the wall clock.

I agree completely.  However it might become possible to remove the need to get wall-clock on every command execution.  This would make the need for wall-clock a more exceptional requirement.  In the SLOWLOG case, we're not really concerned about the extra overhead to call `ustime()` if we've already been handling a slow command.  expiry might have wider implications though.  Ideally, we could execute most commands without ever having to call `ustime()`.