@madolson 2022-04-08
I think Jim's idea is really good. We should be able to start a monotonic clock on every command start. If we run into a case we need a precise time, we can get the local time and then subtract off the time since the start with the monotonic clock. In the case where we don't have a HW clock, we won't do any math, we'll just return the monotonic click value we collected at the start.