@filipecosta90 2022-04-11
@yoav-steinberg WRT to:
> Right. So I think we need to change this PR so that if USE_PROCESSOR_CLOCK is defined we use the HW clock for cmd duration and if it's not defined we use the proposed logic.

using the last commit on this PR and testing with/without HW clock here's the data:
unstable:
```
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets            0.00          ---          ---             ---             ---             ---             ---         0.00 
Gets       676114.65    676114.65         0.00         4.43638         3.74300         6.94300         7.48700    688585.47 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     676114.65    676114.65         0.00         4.43638         3.74300         6.94300         7.48700    688585.47 
```

ustime:
```
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets            0.00          ---          ---             ---             ---             ---             ---         0.00 
Gets       689547.70    689547.70         0.00         4.34994         3.74300         7.29500         7.67900    702266.28 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     689547.70    689547.70         0.00         4.34994         3.74300         7.29500         7.67900    702266.28 
```

hw clock (monotonic clock: X86 TSC @ 2100 ticks/us) :
```
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets            0.00          ---          ---             ---             ---             ---             ---         0.00 
Gets       691869.31    691869.31         0.00         4.33539         3.71100         7.23100         7.58300    704630.70 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     691869.31    691869.31         0.00         4.33539         3.71100         7.23100         7.58300    704630.70
```

this means we get ~2% improvement by default @oranagra. the HW clock reduces the overhead of time tracking by 0.3%. 


a bit of detail on the test system:
```
Linux hpe10 5.4.0-107-generic #121-Ubuntu SMP Thu Mar 24 16:04:27 UTC 2022 x86_64 x86_64 x86_64 GNU/Linux
CPU: Intel(R) Xeon(R) Gold 6230 CPU @ 2.10GHz (hyperthreading disabled)
used compiler: gcc-11 (11.1.0)
```