@oranagra 2022-04-11
In unstable we have 3 time system calls per command, and now in your PR (and in Redis 5.0) we had 2 system calls, this change accounts for 2% performance saving. 
But then when defining HW clock, we have just one system call per command, so why does that provide just 0.3% and not an additional 2%?